+++ Bloxstrap/Models/GameJoinData.cs
@@ -4,7 +4,7 @@
     {
         public GameJoinType JoinType = GameJoinType.Unknown;
 
-        public int? PlaceId { get; set; }
+        public long? PlaceId { get; set; }
         public string? JobId { get; set; }
         public int? UserId { get; set; }
         public string? AccessCode { get; set; }
